Add cljs-asset-path option
The old asset-path option had the opposite behavior of the option of
the same name in the clojurescript compiler.  It was performing a role
similar to clojurescript's output-dir, and there was no way to get the
expected asset-path behaviour.

See https://github.com/clojure/clojurescript/wiki/Compiler-Options

This commit adds a new option cljs-asset-path that corresponds to the
clojurescript asset-path.