Support multiple reloads in a pipeline; helps #41
This commit does two things:

1. Use a gensym'd name for the (former) `adzerk.boot_reload` ns, so that
different `reload` tasks in a single pipeline can operate independently
without collisions.

2. Introduce an optional `:only-by-re` option, and when it's set, only
send reloads of files in the fileset which match the given patterns.
This presents a solution to the problem in #41, in that reloads can be
limited to the artefacts from a single build's compilation artefacts,
avoiding the need for the `reload` task to do dependency tracking.

For example:

    (deftask dev []
      (comp (watch)
            (reload :ids #{"react-www"}
                    :only-by-re [#"^react-www\.out/.*"]
                    :on-jsload 'my.react-www.app/reload)
            (reload :ids #{"react-native"}
                    :only-by-re [#"^react-native\.out/.*"]
                    :on-jsload 'my.react-native.app/reload)
            (cljs :ids #{"react-www" "react-native"}
                  :optimizations :none)))